# Implemented upb_enumvaldef, for storing information about enumvals.