--- a/upb/json_decode.c
+++ b/upb/json_decode.c
@@ -814,10 +814,13 @@ static upb_msgval jsondec_strfield(jsondec *d, const upb_fielddef *f) {
 static upb_msgval jsondec_enum(jsondec *d, const upb_fielddef *f) {
   switch (jsondec_peek(d)) {
     case JD_STRING: {
-      const upb_enumdef *e = upb_fielddef_enumsubdef(f);
       upb_strview str = jsondec_string(d);
+      const upb_enumdef *e = upb_fielddef_enumsubdef(f);
+      const upb_enumvaldef *ev = upb_enumdef_lookupname(e, str.data, str.size);
       upb_msgval val;
-      if (!upb_enumdef_ntoi(e, str.data, str.size, &val.int32_val)) {
+      if (ev) {
+        val.int32_val = upb_enumvaldef_number(ev);
+      } else {
         if (d->options & UPB_JSONDEC_IGNOREUNKNOWN) {
           val.int32_val = 0;
         } else {